+++ tests/testthat/test-ts.R
@@ -121,11 +121,11 @@ test_that('autoplot works for xts', {
   library(xts)
   p <- autoplot(as.xts(AirPassengers), ts.colour = 'green')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
 
   p <- autoplot(as.xts(Canada), ts.colour = 'green')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
 })
 
 test_that('autoplot ts works for univariate timeseries', {
@@ -133,7 +133,7 @@ test_that('autoplot ts works for univariate timeseries', {
 
   p <- autoplot(sts)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 2, 3, 4))
   expect_equal(ld$x, c(1, 2, 3, 4))
@@ -141,7 +141,7 @@ test_that('autoplot ts works for univariate timeseries', {
 
   p <- autoplot(sts, facets = TRUE, stacked = TRUE, colour = 'blue')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 2, 3, 4))
   expect_equal(ld$x, c(1, 2, 3, 4))
@@ -150,7 +150,7 @@ test_that('autoplot ts works for univariate timeseries', {
 
   p <- autoplot(sts, facets = TRUE, stacked = TRUE, colour = 'blue', geom = 'ribbon')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomRibbon' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomRibbon'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$ymin, c(0, 0, 0, 0))
   expect_equal(ld$ymax, c(1, 2, 3, 4))
@@ -160,7 +160,7 @@ test_that('autoplot ts works for univariate timeseries', {
 
   p <- autoplot(sts, geom = 'bar')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomBar' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomBar'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 2, 3, 4))
   expect_equal(ld$x, c(1, 2, 3, 4))
@@ -175,7 +175,7 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'bar')
   expect_equal(length(p$layers), 1)
-  expect_true('GeomBar' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomBar'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 5, 2, 6, 3, 7, 4, 8)) # not stacked
   expect_equal(ld$colour, rep(c('#F8766D',  '#00BFC4'), 4))
@@ -184,7 +184,7 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'bar', stacked = TRUE)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomBar' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomBar'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 6, 2, 8, 3, 10, 4, 12)) # stacked
   expect_equal(ld$colour, rep(c('#F8766D',  '#00BFC4'), 4))
@@ -193,7 +193,7 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'ribbon', stacked = FALSE)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomRibbon' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomRibbon'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$ymin, c(0, 0, 0, 0, 0, 0, 0, 0)) # not stacked
   expect_equal(ld$ymax, c(1, 2, 3, 4, 5, 6, 7, 8)) # not stacked
@@ -205,7 +205,7 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'ribbon', stacked = TRUE)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomRibbon' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomRibbon'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$ymin, c(0, 0, 0, 0, 1, 2, 3, 4)) # stacked
   expect_equal(ld$ymax, c(1, 2, 3, 4, 6, 8, 10, 12)) # stacked
@@ -216,7 +216,7 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'line', stacked = FALSE)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 2, 3, 4, 5, 6, 7, 8)) # not stacked
   expect_equal(ld$x, c(1, 2, 3, 4, 1, 2, 3, 4))
@@ -225,10 +225,13 @@ test_that('autoplot ts works for multivariate timeseries', {
 
   p <- autoplot(mts, facets=FALSE, geom = 'line', stacked = TRUE)
   expect_equal(length(p$layers), 1)
-  expect_true('GeomLine' %in% class(p$layers[[1]]$geom))
+  expect_true(is(p$layers[[1]]$geom, 'GeomLine'))
   ld <- ggplot2:::layer_data(p, 1)
   expect_equal(ld$y, c(1, 2, 3, 4, 6, 8, 10, 12)) # stacked
   expect_equal(ld$x, c(1, 2, 3, 4, 1, 2, 3, 4))
   expect_equal(ld$colour, rep(c('#F8766D',  '#00BFC4'), c(4, 4)))
   expect_true(all(is.na(ld$alpha)))
+
+  # error cases
+  expect_error(autoplot(mts, geom = 'xxx'))
 })